# apps/backend/src/infrastructure/openapi.yaml
openapi: 3.1.0
info:
  title: Jura API
  description: Backend API for the Jura application
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.jura.com
    description: Production server
...
paths:
...
  /api/v1/agents:
    get:
      tags:
        - Agents
      summary: List Agents
      description: List all agents configured for the current user.
      operationId: listAgents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
        "401":
          description: Unauthorized
    post:
      tags:
        - Agents
      summary: Create Agent
      description: Create a new agent configuration.
      operationId: createAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [PRIVATE, PUBLIC]
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        configuration:
          type: object
          properties:
            systemPrompt:
              type: string
            tone:
              type: string
            enableThreads:
              type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required:
        - name
        - type
        - instructions
        - tone
      properties:
        name:
          type: string
        type:
          type: string
          enum: [PRIVATE, PUBLIC]
        instructions:
          type: string
          description: Base system prompt/instructions for the agent
        tone:
          type: string
          description: Voice tone (e.g., Professional, Casual)
        enableThreads:
          type: boolean
          description: Enable persistent chat history (default false)
          default: false
...